refactor: add conftest.py and put fixtures there
